requirements.txt
langchain==0.3.* 
langchain-core==0.3.*
langgraph
langchain-huggingface 
langchain-openai 
langchain-anthropic 
langchain-google-genai 
langchain-google-vertexai 
langchain-ollama 
langchain-mistralai 
langchain-text-splitters
langchain-cohere
langchain-community
langchain-groq
langchain-experimental
langchain-chroma
langsmith
langgraph
huggingface_hub[hf_xet]
faiss-cpu 
qdrant-client 
langchain-neo4j 
arxiv
pymupdf
unstructured[all-docs] 
pymupdf4llm 
cohere 
bitsandbytes 
litellm 
apscheduler 
fastapi 
uvicorn 
gradio 
whisper-openai 
streamlit 
python-dotenv
datasets
jupyter
docarray
ruff
streamlit
rank_bm25
wikipedia
pymupdf
opencv-python
pillow